+++ src/Compilers/CSharp/Portable/Symbols/ConstantValueUtils.cs
@@ -6,6 +6,7 @@ using System.Collections.Immutable;
 using System.Collections.Generic;
 using System.Diagnostics;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.PooledObjects;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
 {
@@ -68,6 +69,23 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
             }
 
             return result;
+        }
+
+        internal static string UnescapeInterpolatedStringLiteral(string s)
+        {
+            var builder = PooledStringBuilder.GetInstance();
+            var stringBuilder = builder.Builder;
+            int formatLength = s.Length;
+            for (int i = 0; i < formatLength; i++)
+            {
+                char c = s[i];
+                stringBuilder.Append(c);
+                if ((c == '{' || c == '}') && (i + 1) < formatLength && s[i + 1] == c)
+                {
+                    i++;
+                }
+            }
+            return builder.ToStringAndFree();
         }
 
         internal static ConstantValue GetAndValidateConstantValue(
